+++ dali/python/nvidia/dali/types.py
@@ -258,6 +258,14 @@ def _is_torch_tensor(value):
 def _is_numpy_array(value):
     return 'numpy.ndarray' in str(type(value))
 
+def _raw_cuda_stream(stream_obj):
+    if hasattr(stream_obj, "cuda_stream"):  # torch
+        return stream_obj.cuda_stream
+    elif hasattr(stream_obj, "ptr"):  # cupy
+        return stream_obj.ptr
+    else:
+        return stream_obj
+
 _cupy_array_type_regex = re.compile('.*cupy\..*\.ndarray.*')
 
 def _is_cupy_array(value):
